requirements.txt
# Core dependencies for the Prometheus project
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.4.0
h5py>=3.3.0
PyYAML>=5.4.0

# Machine learning dependencies
torch>=1.9.0
torchvision>=0.10.0

# Scientific computing
scikit-learn>=1.0.0

# Development and testing
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.6.0
flake8>=3.9.0

# Optional dependencies for enhanced functionality
tqdm>=4.61.0  # Progress bars
seaborn>=0.11.0  # Enhanced plotting